dashboard_api/requirements.txt
asyncpg==0.31.0
fastapi==0.103.1
httpx==0.28.1
loguru==0.7.3
pydantic==2.12.5
python-dotenv==1.0.0
uvicorn==0.23.2
